Process state populations correctly and relocate index.py
